Add redirection in URLs and project documentation
A redirection functionality from the base path to the 'child_list' view was added to the URLs file. Simultaneously, a comprehensive README file detailing the project's description, features, technologies used, structure, and various other aspects was incorporated. Also, a requirements.txt file was created to keep track of necessary dependencies, in this case Django.